Fixed problems with asynPrint formats

+++ asyn/miscellaneous/asynShellCommands.c
@@ -409,7 +409,7 @@ epicsShareFunc int
         pPvt->timeout,&ninp,&eomReason);
     if (status!=asynSuccess) {
        asynPrint(pasynUser, ASYN_TRACE_ERROR,
-                 "Error reading, ninp=%d error %s\n", ninp,pasynUser->errorMessage);
+                 "Error reading, ninp=%lu error %s\n", (unsigned long)ninp,pasynUser->errorMessage);
        return(-1);
     }
     fprintf(stdout,"eomReason 0x%x\n",eomReason);
@@ -444,8 +444,8 @@ epicsShareFunc int
         len, pPvt->timeout,&nout);
     if (status!=asynSuccess) {
        asynPrint(pasynUser, ASYN_TRACE_ERROR,
-                 "Error in asynOctetWrite, nout=%d, len=%d error %s\n",
-                 nout, len,pasynUser->errorMessage);
+                 "Error in asynOctetWrite, nout=%lu, len=%lu error %s\n",
+                 (unsigned long)nout, (unsigned long)len, pasynUser->errorMessage);
        return(-1);
     }
     return (int)nout;
@@ -483,8 +483,8 @@ epicsShareFunc int
                                   &nout,&ninp,&eomReason);
     if (status!=asynSuccess) {
        asynPrint(pasynUser, ASYN_TRACE_ERROR,
-           "Error in WriteRead, nout %d ninp=%d error %s\n",
-           nout, ninp,pasynUser->errorMessage);
+           "Error in WriteRead, nout %lu ninp=%lu error %s\n",
+           (unsigned long)nout, (unsigned long)ninp, pasynUser->errorMessage);
        return(-1);
     }
     fprintf(stdout,"eomReason 0x%x\n",eomReason);
